controle pour suppression

+++ project/test/unit/giilda99SuppressionSocietesComptesEtablissementsTest.php
@@ -16,16 +16,18 @@ foreach (CompteTagsView::getInstance()->listByTags('test', 'test') as $k => $v) 
     if (preg_match('/SOCIETE-([^ ]*)/', implode(' ', array_values($v->value)), $m)) {
         $soc = SocieteClient::getInstance()->findByIdentifiantSociete($m[1]);
         foreach($soc->getEtablissementsObj() as $k => $etabl) {
-            foreach (VracClient::getInstance()->retrieveBySoussigne($etabl->etablissement->identifiant)->rows as $k => $vrac) {
-                $vrac_obj = VracClient::getInstance()->find($vrac->id);
-                $vrac_obj->delete();
-                $t->is(VracClient::getInstance()->find($vrac->id), null, "Suppression du contrat ".$vrac->id);
-            }
-            foreach (DRMClient::getInstance()->viewByIdentifiant($etabl->etablissement->identifiant) as $id => $drm) {
-                $drm = DRMClient::getInstance()->find($id);
-                $drm->delete(false);
-                $t->is(DRMClient::getInstance()->find($id), null, "Suppression de la DRM ".$id);
-            }
+            if($etabl->etablissement){
+              foreach (VracClient::getInstance()->retrieveBySoussigne($etabl->etablissement->identifiant)->rows as $k => $vrac) {
+                  $vrac_obj = VracClient::getInstance()->find($vrac->id);
+                  $vrac_obj->delete();
+                  $t->is(VracClient::getInstance()->find($vrac->id), null, "Suppression du contrat ".$vrac->id);
+              }
+              foreach (DRMClient::getInstance()->viewByIdentifiant($etabl->etablissement->identifiant) as $id => $drm) {
+                  $drm = DRMClient::getInstance()->find($id);
+                  $drm->delete(false);
+                  $t->is(DRMClient::getInstance()->find($id), null, "Suppression de la DRM ".$id);
+              }
+          }
         }
         foreach (FactureSocieteView::getInstance()->findBySociete($soc) as $id => $facture) {
             $facture = FactureClient::getInstance()->find($id);
